# qrios-ussd-api-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QRIOS USSD API documentation",
    "description": "This API is *not* exposed by Qrios, but it should be implemented by the developer. Qrios will call this API on developer's side to provide information about USSD sessions and the developer should answer with information regarding what should happen next in the aforementioned sessions.",
    "version": "1.6.0"
  },
  "servers": [
    {
      "url": "https://example-developer-ussd-app.com",
      "description": "URL to developer's USSD app"
    }
  ],
...
  "paths": {
...
    "/ussdSessionEvent/close": {
      "post": {
        "tags": [
          "Developer's app endpoints"
        ],
        "summary": "Called when a USSD session is closing gracefully (user cannot provide more input, user exits, session times out)",
        "description": "This endpoint is called when the session closes gracefully, i.e. the end of the session is not caused by an error. There are three ways in which that can happen - the session may end naturally, as no input can be provided by the user (that happens when InfoView is used); the user may end the session by pressing \"Cancel\" on their phone; or the session is ended by the mobile operator due to timeout (e.g. two minutes).<br/>No response body is expected at this point as the session has been closed.",
        "operationId": "postUssdsessioneventClose",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "The authorization secret value according to the configured authorization method of the USSD app",
            "required": false,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseSession"
              },
              "examples": {
                "CloseEndExample": {
                  "summary": "Session ends naturally",
                  "description": "The session closes, as further input is not expected.",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "End"
                    },
                    "contextData": "devCorrelationId:12345"
                  }
                },
                "CloseAbandonExample": {
                  "summary": "Session is abandoned",
                  "description": "The session closes, as the user has left it (e.g. by pressing Cancel on their phone).",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "Abandon"
                    },
                    "contextData": "devCorrelationId:12345"
                  }
                },
                "CloseTimeoutExample": {
                  "summary": "Session times out",
                  "description": "The session is closed by mobile operator due to being too long.",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "Timeout"
                    },
                    "contextData": "devCorrelationId:12345"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session end has been successfully handled by the developer; the server acknowledges that the session has been closed. Response body is empty."
          },
          "204": {
            "description": "Session end has been successfully handled by the developer; the server acknowledges that the session has been closed. Response body is empty."
          }
        }
      }
    },
    "/ussdSessionEvent/abort": {
      "post": {
        "tags": [
          "Developer's app endpoints"
        ],
        "summary": "Called when a USSD session is being aborted (internal error)",
        "description": "This endpoint is called when the session is aborted due to an internal error. This can happen in a number of ways, which are not usually detailed - the exception being an error of insufficient funds available in the virtual purse. When such an error happens, the USSD session is ended and a request is sent to this endpoint to inform of an error. <br/>No response body is expected at this point as the session has been aborted.",
        "operationId": "postUssdsessioneventAbort",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "The authorization secret value according to the configured authorization method of the USSD app",
            "required": false,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortSession"
              },
              "examples": {
                "CloseEndExample": {
                  "summary": "Missing Privilege",
                  "description": "The session is aborted due to missing USSD privilege",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "privilege": "UssdApi",
                      "type": "MissingPrivilege"
                    }
                  }
                },
                "AbortInsufficientBalanceExample": {
                  "summary": "Insufficient funds",
                  "description": "The session is aborted as you don't have enough funds to continue using the service. Top up your virtual purse to continue using the service",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "InsufficientBalanceInVirtualPurse"
                    }
                  }
                },
                "AbortDuplicatedOperationExample": {
                  "summary": "Duplicated Operation",
                  "description": "The session is aborted because of trying to invoke the same operation (like merchant payment) once again",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "operationId": "535bdec5-ea1c-4737-b9e2-04a9743d4227",
                      "type": "DuplicatedOperation"
                    }
                  }
                },
                "AbortUnexpectedUssdAppResponseExample": {
                  "summary": "Unexpected USSD App Response",
                  "description": "The session is aborted because previous USSD app response was malformed or not expected",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "UnexpectedUssdAppResponse"
                    }
                  }
                },
                "AbortInternalErrorExample": {
                  "summary": "Internal Error",
                  "description": "The session is aborted due to internal error in Qrios API",
                  "value": {
                    "clientId": "x97uJ23b",
                    "appId": "AppIdGeneratedByQriosApi",
                    "sessionId": "8a444af8-4b03-4545-8677-2a6c924bdddb",
                    "reason": {
                      "type": "InternalError"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The aborting of the session has been successfully handled by the developer; the server acknowledges that the session has been aborted. Response body is empty."
          },
          "204": {
            "description": "The aborting of the session has been successfully handled by the developer; the server acknowledges that the session has been aborted. Response body is empty."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AbortSession": {
        "required": [
          "appId",
          "clientId",
          "reason",
          "sessionId"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Client identifier of the developer using Qrios platform."
          },
          "appId": {
            "type": "string",
            "description": "Identifier of the USSD app created on Qrios platform.",
            "example": "b56r455"
          },
          "sessionId": {
            "type": "string",
            "description": "Unique identifier of USSD session."
          },
          "reason": {
            "$ref": "#/components/schemas/AbortSession.AbortReason"
          }
        },
        "description": "The event sent when the USSD session is aborted (internal error)."
      },
      "AbortSession.AbortReason": {
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "DuplicatedOperation": "#/components/schemas/AbortSession.AbortReason.DuplicatedOperation",
            "InsufficientBalanceInVirtualPurse": "#/components/schemas/AbortSession.AbortReason.InsufficientBalanceInVirtualPurse",
            "InternalError": "#/components/schemas/AbortSession.AbortReason.InternalError",
            "MissingPrivilege": "#/components/schemas/AbortSession.AbortReason.MissingPrivilege",
            "UnexpectedUssdAppResponse": "#/components/schemas/AbortSession.AbortReason.UnexpectedUssdAppResponse"
          }
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AbortSession.AbortReason.DuplicatedOperation"
          },
          {
            "$ref": "#/components/schemas/AbortSession.AbortReason.InsufficientBalanceInVirtualPurse"
          },
          {
            "$ref": "#/components/schemas/AbortSession.AbortReason.InternalError"
          },
          {
            "$ref": "#/components/schemas/AbortSession.AbortReason.MissingPrivilege"
          },
          {
            "$ref": "#/components/schemas/AbortSession.AbortReason.UnexpectedUssdAppResponse"
          }
        ]
      },
      "AbortSession.AbortReason.DuplicatedOperation": {
        "required": [
          "operationId",
          "type"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "description": "USSD process abort reason - operation, eg. Merchant Payment, was previously invoked with given identifier."
      },
...
      "CloseSession": {
        "required": [
          "appId",
          "clientId",
          "contextData",
          "reason",
          "sessionId"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Client identifier of the developer using Qrios platform."
          },
          "appId": {
            "type": "string",
            "description": "Identifier of the USSD app created on Qrios platform.",
            "example": "b56r455"
          },
          "sessionId": {
            "type": "string",
            "description": "Unique identifier of USSD session."
          },
          "reason": {
            "$ref": "#/components/schemas/CloseSession.CloseReason"
          },
          "contextData": {
            "type": "string",
            "description": "String set by the developer and carried over throughout the session."
          }
        },
        "description": "The event sent when the USSD session is closed (user can no longer input data, user leaves the session, session times out)."
      },
      "CloseSession.CloseReason": {
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "Abandon": "#/components/schemas/CloseSession.CloseReason.Abandon",
            "End": "#/components/schemas/CloseSession.CloseReason.End",
            "Timeout": "#/components/schemas/CloseSession.CloseReason.Timeout"
          }
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/CloseSession.CloseReason.Abandon"
          },
          {
            "$ref": "#/components/schemas/CloseSession.CloseReason.End"
          },
          {
            "$ref": "#/components/schemas/CloseSession.CloseReason.Timeout"
          }
        ]
      },
      "CloseSession.CloseReason.Abandon": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          }
        },
        "description": "Abandon - the user leaves the session (e.g. the user presses \"Cancel\" on the phone)."
      },
      "CloseSession.CloseReason.End": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          }
        },
        "description": "End - the session ends naturally (e.g. the user can no longer input data)."
      },
      "CloseSession.CloseReason.Timeout": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          }
        },
        "description": "Timeout - the session is ended by the mobile operator (e.g. after two minutes)."
      },
...
    }
  }
}